performance fixes
i had  todo some performance fixes on the code, because it was down trash. problably because i was doing vector allocations with the charset pool.

to counter this, i now use a single ''constexpr array<char, 68>'' which holds the entire charset i use on this program.
the problem is that now i cant have custom error messages(rather, i dont know how to), depending on which part of the charset the program fails to find a match for the password characters.

ill eventually figure out a cleaner way to fix this, but for now this works fine.